+++ desc/objectives/getters.py
@@ -22,6 +22,9 @@ from .linear_objectives import (
     FixIonTemperature,
     FixIota,
     FixLambdaGauge,
+    FixNearAxisLambda,
+    FixNearAxisR,
+    FixNearAxisZ,
     FixPressure,
     FixPsi,
     FixSheetCurrent,
@@ -147,9 +150,10 @@ def get_NAE_constraints(
     desc_eq : Equilibrium
         Equilibrium to constrain behavior of
         (assumed to be a fit from the NAE equil using `.from_near_axis()`).
-    qsc_eq : Qsc
+    qsc_eq : Qsc, optional
         Qsc object defining the near-axis equilibrium to constrain behavior to.
-    order : int
+        if None, will instead fix the current near-axis behavior of the ``desc_eq``
+    order : {0,1,2}
         order (in rho) of near-axis behavior to constrain
     profiles : bool
         If True, also include constraints to fix all profiles assigned to equilibrium.
@@ -172,7 +176,7 @@ def get_NAE_constraints(
     kwargs = {"eq": desc_eq, "normalize": normalize, "normalize_target": normalize}
     if not isinstance(fix_lambda, bool):
         fix_lambda = int(fix_lambda)
-    constraints = (FixAxisR(**kwargs), FixAxisZ(**kwargs), FixPsi(**kwargs))
+    constraints = (FixPsi(**kwargs),)
 
     if profiles:
         for name, con in _PROFILE_CONSTRAINTS.items():
@@ -180,11 +184,102 @@ def get_NAE_constraints(
                 constraints += (con(**kwargs),)
     constraints += (FixSheetCurrent(**kwargs),)
 
+    constraints += (
+        FixNearAxisR(
+            eq=desc_eq,
+            target=qsc_eq,
+            N=N,
+            order=order,
+            normalize=normalize,
+        ),
+        FixNearAxisZ(
+            eq=desc_eq,
+            target=qsc_eq,
+            N=N,
+            order=order,
+            normalize=normalize,
+        ),
+    )
+    if fix_lambda or (fix_lambda >= 0 and type(fix_lambda) is int):
+        constraints += (
+            FixNearAxisLambda(
+                eq=desc_eq,
+                target=qsc_eq,
+                N=N,
+                order=int(fix_lambda),
+                normalize=normalize,
+            ),
+        )
+
+    return constraints
+
+
+def _get_NAE_constraints(
+    desc_eq,
+    qsc_eq,
+    order=1,
+    N=None,
+    fix_lambda=False,
+    normalize=True,
+):
+    """Get the constraints necessary for fixing NAE behavior in an equilibrium problem.
+
+    NOTE: This will return tuples of FixSumModes__, this is not intended to be directly
+    used by the user. Instead, call the ``get_NAE_constraints`` function, or use the
+    FixNearAxis{R,Z,Lambda} objectives along with the FixAxis{R,Z} objectives. This
+    instead is a helper function to get the needed constraints for the
+    FixNearAxis{R,Z,Lambda} objectives, which use those constraints to form
+    the full constraint matrices to properly constrain the behavior of that
+    part of the equilibrium.
+
+    Parameters
+    ----------
+    desc_eq : Equilibrium
+        Equilibrium to constrain behavior of
+        (assumed to be a fit from the NAE equil using `.from_near_axis()`).
+    qsc_eq : Qsc
+        Qsc object defining the near-axis equilibrium to constrain behavior to.
+        order : {0,1,2}
+        order (in rho) of near-axis behavior to constrain
+    normalize : bool
+        Whether to apply constraints in normalized units.
+    N : int
+        max toroidal resolution to constrain.
+        If `None`, defaults to equilibrium's toroidal resolution
+    fix_lambda : bool or int
+        Whether to constrain lambda to match that of the NAE near-axis
+        if an `int`, fixes lambda up to that order in rho {0,1}
+        if `True`, fixes lambda up to the specified order given by `order`
+    normalize : bool
+        Whether to apply constraints in normalized units.
+
+    Returns
+    -------
+    constraints, tuple of _Objectives
+        A list of the linear constraints used in fixed-near-axis behavior problems.
+    """
+    if not isinstance(fix_lambda, bool):
+        fix_lambda = int(fix_lambda)
+    constraints = ()
+
+    kwargs = {"eq": desc_eq, "normalize": normalize, "normalize_target": normalize}
+    if not isinstance(fix_lambda, bool):
+        fix_lambda = int(fix_lambda)
+
     if fix_lambda or (fix_lambda >= 0 and type(fix_lambda) is int):
         L_axis_constraints, _, _ = calc_zeroth_order_lambda(
             qsc=qsc_eq, desc_eq=desc_eq, N=N
         )
         constraints += L_axis_constraints
+
+    # Axis constraints
+    constraints += (
+        FixAxisR(**kwargs),
+        FixAxisZ(**kwargs),
+        AxisRSelfConsistency(desc_eq),
+        AxisZSelfConsistency(desc_eq),
+    )
+
     if order >= 1:  # first order constraints
         constraints += make_RZ_cons_1st_order(
             qsc=qsc_eq, desc_eq=desc_eq, N=N, fix_lambda=fix_lambda and fix_lambda > 0