+++ desc/magnetic_fields/_current_potential.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from desc.backend import fori_loop, jnp
 from desc.basis import DoubleFourierSeries
-from desc.compute import rpz2xyz, rpz2xyz_vec, xyz2rpz, xyz2rpz_vec
+from desc.compute import rpz2xyz, rpz2xyz_vec, xyz2rpz_vec
 from desc.compute.utils import _compute as compute_fun
 from desc.geometry import FourierRZToroidalSurface
 from desc.grid import LinearGrid
@@ -795,7 +795,7 @@ def _compute_A_or_B_from_CurrentPotentialField(
         data = field.compute(
             ["K", "x"],
             grid=source_grid,
-            basis="xyz",
+            basis="rpz",
             params=params,
             transforms=transforms,
             jitable=True,
@@ -807,11 +807,10 @@ def _compute_A_or_B_from_CurrentPotentialField(
             params=params,
             transforms=transforms,
             profiles={},
-            basis="xyz",
         )
 
-    _rs = xyz2rpz(data["x"])
-    _K = xyz2rpz_vec(data["K"], phi=source_grid.nodes[:, 2])
+    _rs = data["x"]
+    _K = data["K"]
 
     # surface element, must divide by NFP to remove the NFP multiple on the
     # surface grid weights, as we account for that when doing the for loop
@@ -840,175 +839,3 @@ def _compute_A_or_B_from_CurrentPotentialField(
     if basis == "rpz":
         B = xyz2rpz_vec(B, x=coords[:, 0], y=coords[:, 1])
     return B
-
-
-def _compute_magnetic_field_from_CurrentPotentialField(
-    field, coords, source_grid, params=None, basis="rpz", transforms=None
-):
-    """Compute magnetic field at a set of points.
-
-    Parameters
-    ----------
-    field : CurrentPotentialField or FourierCurrentPotentialField
-        current potential field object from which to compute magnetic field.
-    coords : array-like shape(N,3)
-        cylindrical or cartesian coordinates
-    source_grid : Grid,
-        source grid upon which to evaluate the surface current density K
-    params : dict, optional
-        parameters to pass to compute function
-        should include the potential
-    basis : {"rpz", "xyz"}
-        basis for input coordinates and returned magnetic field
-
-
-    Returns
-    -------
-    field : ndarray, shape(N,3)
-        magnetic field at specified points
-
-    """
-    assert basis.lower() in ["rpz", "xyz"]
-    coords = jnp.atleast_2d(jnp.asarray(coords))
-    if basis == "rpz":
-        coords = rpz2xyz(coords)
-
-    # compute surface current, and store grid quantities
-    # needed for integration in class
-    # TODO: does this have to be xyz, or can it be computed in rpz as well?
-    if not params or not transforms:
-        data = field.compute(
-            ["K", "x"],
-            grid=source_grid,
-            basis="xyz",
-            params=params,
-            transforms=transforms,
-            jitable=True,
-        )
-    else:
-        data = compute_fun(
-            field,
-            names=["K", "x"],
-            params=params,
-            transforms=transforms,
-            profiles={},
-            basis="xyz",
-        )
-
-    _rs = xyz2rpz(data["x"])
-    _K = xyz2rpz_vec(data["K"], phi=source_grid.nodes[:, 2])
-
-    # surface element, must divide by NFP to remove the NFP multiple on the
-    # surface grid weights, as we account for that when doing the for loop
-    # over NFP
-    _dV = source_grid.weights * data["|e_theta x e_zeta|"] / source_grid.NFP
-
-    def nfp_loop(j, f):
-        # calculate (by rotating) rs, rs_t, rz_t
-        phi = (source_grid.nodes[:, 2] + j * 2 * jnp.pi / source_grid.NFP) % (
-            2 * jnp.pi
-        )
-        # new coords are just old R,Z at a new phi (bc of discrete NFP symmetry)
-        rs = jnp.vstack((_rs[:, 0], phi, _rs[:, 2])).T
-        rs = rpz2xyz(rs)
-        K = rpz2xyz_vec(_K, phi=phi)
-        fj = biot_savart_general(
-            coords,
-            rs,
-            K,
-            _dV,
-        )
-        f += fj
-        return f
-
-    B = fori_loop(0, source_grid.NFP, nfp_loop, jnp.zeros_like(coords))
-    if basis == "rpz":
-        B = xyz2rpz_vec(B, x=coords[:, 0], y=coords[:, 1])
-    return B
-
-
-def _compute_vector_potential_from_CurrentPotentialField(
-    field, coords, source_grid, params=None, basis="rpz", transforms=None
-):
-    """Compute magnetic vector potential at a set of points.
-
-    Parameters
-    ----------
-    field : CurrentPotentialField or FourierCurrentPotentialField
-        current potential field object from which to compute magnetic vector potential.
-    coords : array-like shape(N,3)
-        cylindrical or cartesian coordinates
-    source_grid : Grid,
-        source grid upon which to evaluate the surface current density K
-    params : dict, optional
-        parameters to pass to compute function
-        should include the potential
-    basis : {"rpz", "xyz"}
-        basis for input coordinates and returned magnetic vector potential
-    transforms : dict of Transform
-            Transforms for R, Z, lambda, etc. Default is to build from source_grid
-
-
-    Returns
-    -------
-    A : ndarray, shape(N,3)
-        magnetic vector potential at specified points
-
-    """
-    assert basis.lower() in ["rpz", "xyz"]
-    coords = jnp.atleast_2d(jnp.asarray(coords))
-    if basis == "rpz":
-        coords = rpz2xyz(coords)
-
-    # compute surface current, and store grid quantities
-    # needed for integration in class
-    # TODO: does this have to be xyz, or can it be computed in rpz as well?
-    if not params or not transforms:
-        data = field.compute(
-            ["K", "x"],
-            grid=source_grid,
-            basis="xyz",
-            params=params,
-            transforms=transforms,
-            jitable=True,
-        )
-    else:
-        data = compute_fun(
-            field,
-            names=["K", "x"],
-            params=params,
-            transforms=transforms,
-            profiles={},
-            basis="xyz",
-        )
-
-    _rs = xyz2rpz(data["x"])
-    _K = xyz2rpz_vec(data["K"], phi=source_grid.nodes[:, 2])
-
-    # surface element, must divide by NFP to remove the NFP multiple on the
-    # surface grid weights, as we account for that when doing the for loop
-    # over NFP
-    _dV = source_grid.weights * data["|e_theta x e_zeta|"] / source_grid.NFP
-
-    def nfp_loop(j, f):
-        # calculate (by rotating) rs, rs_t, rz_t
-        phi = (source_grid.nodes[:, 2] + j * 2 * jnp.pi / source_grid.NFP) % (
-            2 * jnp.pi
-        )
-        # new coords are just old R,Z at a new phi (bc of discrete NFP symmetry)
-        rs = jnp.vstack((_rs[:, 0], phi, _rs[:, 2])).T
-        rs = rpz2xyz(rs)
-        K = rpz2xyz_vec(_K, phi=phi)
-        fj = biot_savart_general_vector_potential(
-            coords,
-            rs,
-            K,
-            _dV,
-        )
-        f += fj
-        return f
-
-    A = fori_loop(0, source_grid.NFP, nfp_loop, jnp.zeros_like(coords))
-    if basis == "rpz":
-        A = xyz2rpz_vec(A, x=coords[:, 0], y=coords[:, 1])
-    return A