+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+New Features
+
+- adds ``from_values`` method that was present in ``FourierRZCurve`` but missing in ``FourierRZCoil``
+- also adds new ``from_values`` method for ``FourierPlanarCurve`` and ``FourierPlanarCoil``
+
 
 v0.12.0
 -------